console: Refactor connect in brand select
